+++ CHANGELOG.md
@@ -7,6 +7,7 @@
     * The Image svelte component is now at the `/Image.svelte` entry point (`import Image from 'svimg/Image.svelte'` instead of `import Image from 'svelte'`)
     * The s-image custom element is now at the `/s-image` entry point (`import 'svimg/s-image'` instead of `import 'svimg/dist/s-image'`)
     * The process entry point's contents are now just part of the default `'svimg'` entry point
+* BREAKING: Remove `publicPath` option. Please use a `srcGenerator` function instead: `(path) => '/public/path' + path`
 * BREAKING: Drop Node 14 support
 * BREAKING: Target ES2021. Support modern browsers and Node versions, but not older browsers like IE